[IMPROVE] improve feature customer class

+++ src/main/java/com/enigma/tokopakedi/service/impl/CustomerServiceImpl.java
@@ -5,11 +5,14 @@ import com.enigma.tokopakedi.entity.UserCredential;
 import com.enigma.tokopakedi.model.reponse.CustomerResponse;
 import com.enigma.tokopakedi.model.reponse.RoleResponse;
 import com.enigma.tokopakedi.model.reponse.UserResponse;
+import com.enigma.tokopakedi.model.request.CustomerRequest;
 import com.enigma.tokopakedi.model.request.SearchCustomerRequest;
 import com.enigma.tokopakedi.model.reponse.UserCredentialResponse;
 import com.enigma.tokopakedi.repository.CustomerRepository;
 import com.enigma.tokopakedi.repository.UserCredentialRepository;
 import com.enigma.tokopakedi.service.CustomerService;
+import com.enigma.tokopakedi.service.UserService;
+import jakarta.persistence.RollbackException;
 import jakarta.persistence.criteria.Predicate;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,6 +22,7 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
@@ -31,9 +35,11 @@ import java.util.Optional;
 public class CustomerServiceImpl implements CustomerService {
 
     private final CustomerRepository customerRepository;
+    private final UserService userService;
 
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public CustomerResponse readIdByCustomer(String customerId) {
         Customer customers = new Customer();
         customers.setId(customerId);
@@ -60,12 +66,14 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public Customer createNewCustomer(String customerId, Customer customer) {
         customer.setId(customerId);
         return customerRepository.save(customer);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public List<CustomerResponse> getCustomer() {
         List<Customer> customer =customerRepository.findAll();
 
@@ -97,15 +105,19 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public Customer createCustomer(Customer customer) {
         return customerRepository.save(customer);
     }
 
     @Override
-    public void deleteByIdCustomer(String customerId){
-        Customer customer = new Customer();
-        customer.setId(customerId);
+    @Transactional(rollbackFor = Exception.class)
+    public String deleteByIdCustomer(String customerId){
+        Optional<Customer> byId = customerRepository.findById(customerId);
+        if (byId.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        userService.delete(byId.get().getUserCredential().getId());
         customerRepository.deleteById(customerId);
+        return "Data customer deleted";
     }
 
     public Page<Customer> findAllWithPagination(int page, int size) {
@@ -114,6 +126,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public Page<Customer> findAllWithParamm(SearchCustomerRequest request) {
         if (request.getPage()<=0)request.setPage(1);
         PageRequest pageablee = PageRequest.of(request.getPage()-1, request.getSize());
@@ -137,22 +150,26 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public CustomerResponse updateCustomer(Customer customer) {
+    @Transactional(rollbackFor = Exception.class)
+    public CustomerResponse updateCustomer(CustomerRequest customer) {
         Optional<Customer> optionalCustomer = customerRepository.findById(customer.getId());
         if (optionalCustomer.isEmpty()) throw new RuntimeException("customer not found");
 
         List<UserCredentialResponse> userCredentialResponses = new ArrayList<>();
 
         UserCredential currentUserCredential = (UserCredential) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        UserCredential credential = optionalCustomer.get().getUserCredential();
-        if (!currentUserCredential.getId().equals(credential.getId())) throw new ResponseStatusException(HttpStatus.FORBIDDEN, "forbidden");
-        customer.setUserCredential(credential);
-        Customer updatedCustomer = customerRepository.save(customer);
-//        UserCredentialRequest userCredentialRequest = UserCredentialRequest.builder()
-//                .id(optionalCustomer.get().getUserCredential().getId())
-//                .roles(optionalCustomer.get().getUserCredential().getRoles())
-//                .build();
-//        Optional<UserCredential> credential1 = userCredentialRepository.findById(userCredentialRequest.getId());
+        UserCredential userCredential = userService.loadByUserId(customer.getUserCredential());
+        if (!currentUserCredential.getId().equals(userCredential.getId())) throw new ResponseStatusException(HttpStatus.FORBIDDEN, "forbidden");
+        Customer customer1 = Customer.builder()
+                .id(customer.getId())
+                .name(customer.getName())
+                .phone(customer.getPhone())
+                .address(customer.getAddress())
+                .poin(customer.getPoin())
+                .userCredential(userCredential)
+                .build();
+
+        Customer updatedCustomer = customerRepository.save(customer1);
         UserCredentialResponse credentialResponse = UserCredentialResponse.builder()
                 .id(updatedCustomer.getUserCredential().getId())
                 .email(updatedCustomer.getUserCredential().getEmail())
@@ -165,6 +182,7 @@ public class CustomerServiceImpl implements CustomerService {
                 .name(updatedCustomer.getName())
                 .address(updatedCustomer.getAddress())
                 .phone(updatedCustomer.getPhone())
+                .poin(updatedCustomer.getPoin())
                 .userCredential(userCredentialResponses)
                 .build();
         return customerResponse;
